+++ packages/jest-resolve/src/defaultResolver.ts
@@ -37,7 +37,7 @@ export type PackageFilter = (
 ) => PackageJSON;
 
 /**
- * Allows transforms a path within a package.
+ * Allows transforming a path within a package.
  *
  * @param pkg - Parsed `package.json` contents.
  * @param path - Path being resolved.
@@ -51,7 +51,7 @@ export type PathFilter = (
   relativePath: string,
 ) => string;
 
-type ResolverOptions = {
+export type ResolverOptions = {
   /** Directory to begin resolving from. */
   basedir: string;
   /** List of export conditions. */